feat(check_toc_txt): add a script lint toc.txt
Add a script to find invalid / outdated toc.txt entries. This replaces
the older, faster, but more sloppy check-configs.sh script.

I say the old version was more sloppy as it didn't treat comments quite
the same as the rest of the tools here. It was a quick and dirty check.

Signed-off-by: Randolph Sapp <rs@ti.com>